fix: moved all functions and types from facilityUtils.ts to facilitiesStore.ts
